+++ packages/wonder-blocks-link/src/components/link.tsx
@@ -4,8 +4,8 @@ import {Link as ReactRouterLink} from "react-router-dom";
 import {getClickableBehavior} from "@khanacademy/wonder-blocks-clickable";
 
 import type {AriaProps, StyleType} from "@khanacademy/wonder-blocks-core";
+import Icon from "@khanacademy/wonder-blocks-icon";
 import type {Typography} from "@khanacademy/wonder-blocks-typography";
-import type {IconAsset} from "@khanacademy/wonder-blocks-icon";
 import LinkCore from "./link-core";
 
 type CommonProps = AriaProps & {
@@ -123,13 +123,13 @@ type CommonProps = AriaProps & {
     /**
      * An optional icon displayed before the link label.
      */
-    startIcon?: IconAsset;
+    startIcon?: React.ReactElement<typeof Icon>;
     /**
      * An optional icon displayed after the link label.
      * If `target="_blank"` and `endIcon` is passed in, `endIcon` will override
      * the default `externalIcon`.
      */
-    endIcon?: IconAsset;
+    endIcon?: React.ReactElement<typeof Icon>;
 };
 
 export type SharedProps =