+++ packages/wonder-blocks-switch/src/index.ts
@@ -1,0 +1,3 @@
+import Switch from "./components/switch";
+
+export default Switch;