+++ packages/wonder-blocks-theming/src/index.ts
@@ -1,2 +1,12 @@
-// Just a temporary export to include some code.
-export const NOTHING = "NOTHING";
+export {default as tokens} from "./tokens";
+export {mergeTheme} from "./utils/merge-theme";
+export {default as createThemeContext} from "./utils/create-theme-context";
+export {default as useScopedTheme} from "./hooks/use-scoped-theme";
+// HOC
+export {
+    default as withScopedTheme,
+    type WithThemeProps,
+} from "./components/with-scoped-theme";
+export {type ThemedStylesFn} from "./types";
+export {default as useStyles} from "./hooks/use-styles";
+export {ThemeSwitcherContext} from "./utils/theme-switcher-context";