+++ packages/wonder-blocks-theming/src/types.ts
@@ -1,0 +1,3 @@
+import {StyleDeclaration} from "aphrodite";
+
+export type ThemedStylesFn<T> = (theme: T) => StyleDeclaration;