+++ packages/wonder-blocks-tooltip/src/components/tooltip-content.tsx
@@ -5,8 +5,9 @@ import {View} from "@khanacademy/wonder-blocks-core";
 import {Strut} from "@khanacademy/wonder-blocks-layout";
 import Spacing from "@khanacademy/wonder-blocks-spacing";
 import {HeadingSmall, LabelMedium} from "@khanacademy/wonder-blocks-typography";
-
 import type {Typography} from "@khanacademy/wonder-blocks-typography";
+
+import {ContentStyle} from "../util/types";
 
 type Props = {
     /**
@@ -21,6 +22,14 @@ type Props = {
         | string
         | React.ReactElement<React.ComponentProps<Typography>>
         | Array<React.ReactElement<React.ComponentProps<Typography>>>;
+    /**
+     * Optional custom styles for the tooltip which are a subset of valid CSS styles
+     */
+    contentStyle?: ContentStyle;
+    /**
+     * Test ID used for e2e testing.
+     */
+    testId?: string;
 };
 
 /**
@@ -64,7 +73,10 @@ export default class TooltipContent extends React.Component<Props> {
         const children = this._renderChildren();
         const containerStyle = title ? styles.withTitle : styles.withoutTitle;
         return (
-            <View style={containerStyle}>
+            <View
+                style={[containerStyle, this.props.contentStyle]}
+                testId={this.props.testId}
+            >
                 {title}
                 {title && children && <Strut size={Spacing.xxxSmall_4} />}
                 {children}