+++ packages/wonder-blocks-typography/src/components/footnote.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import {Text} from "@khanacademy/wonder-blocks-core";
+import {PropsFor, Text} from "@khanacademy/wonder-blocks-core";
 
 import styles from "../util/styles";
 
-import type {Props} from "../util/types";
+type Props = PropsFor<typeof Text>;
 
 const Footnote = React.forwardRef(function Footnote(
     {style, children, tag = "span", ...otherProps}: Props,