+++ packages/wonder-blocks-birthday-picker/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @khanacademy/wonder-blocks-birthday-picker
 
+## 2.0.21
+
+### Patch Changes
+
+-   Updated dependencies [ec6a33a4]
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-icon@2.1.0
+    -   @khanacademy/wonder-blocks-core@5.4.0
+    -   @khanacademy/wonder-blocks-dropdown@3.0.20
+    -   @khanacademy/wonder-blocks-layout@2.0.16
+    -   @khanacademy/wonder-blocks-typography@2.1.2
+
+## 2.0.20
+
+### Patch Changes
+
+-   Updated dependencies [9f3752d4]
+    -   @khanacademy/wonder-blocks-typography@2.1.1
+    -   @khanacademy/wonder-blocks-core@5.3.1
+    -   @khanacademy/wonder-blocks-dropdown@3.0.19
+    -   @khanacademy/wonder-blocks-icon@2.0.15
+    -   @khanacademy/wonder-blocks-layout@2.0.15
+
 ## 2.0.19
 
 ### Patch Changes