+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @khanacademy/wonder-blocks-link
 
+## 4.4.0
+
+### Minor Changes
+
+-   ec6a33a4: Forward refs in Icon
+    Use Icon type instead of IconAsset for startIcon and endIcon in Link
+
+### Patch Changes
+
+-   Updated dependencies [ec6a33a4]
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-icon@2.1.0
+    -   @khanacademy/wonder-blocks-core@5.4.0
+    -   @khanacademy/wonder-blocks-clickable@3.1.3
+
+## 4.3.1
+
+### Patch Changes
+
+-   9f3752d4: Used named functions in componenets with forwarded refs
+-   Updated dependencies [9f3752d4]
+    -   @khanacademy/wonder-blocks-core@5.3.1
+    -   @khanacademy/wonder-blocks-clickable@3.1.2
+    -   @khanacademy/wonder-blocks-icon@2.0.15
+
 ## 4.3.0
 
 ### Minor Changes