+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @khanacademy/wonder-blocks-popover
 
+## 2.1.7
+
+### Patch Changes
+
+-   Updated dependencies [ec6a33a4]
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-icon@2.1.0
+    -   @khanacademy/wonder-blocks-core@5.4.0
+    -   @khanacademy/wonder-blocks-icon-button@4.1.2
+    -   @khanacademy/wonder-blocks-modal@4.0.18
+    -   @khanacademy/wonder-blocks-tooltip@2.1.1
+    -   @khanacademy/wonder-blocks-typography@2.1.2
+
+## 2.1.6
+
+### Patch Changes
+
+-   Updated dependencies [9f3752d4]
+-   Updated dependencies [9f3752d4]
+    -   @khanacademy/wonder-blocks-icon-button@4.1.1
+    -   @khanacademy/wonder-blocks-typography@2.1.1
+    -   @khanacademy/wonder-blocks-core@5.3.1
+    -   @khanacademy/wonder-blocks-tooltip@2.1.0
+    -   @khanacademy/wonder-blocks-modal@4.0.17
+    -   @khanacademy/wonder-blocks-icon@2.0.15
+
 ## 2.1.5
 
 ### Patch Changes