+++ packages/wonder-blocks-progress-spinner/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-progress-spinner
 
+## 2.0.16
+
+### Patch Changes
+
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-core@5.4.0
+
+## 2.0.15
+
+### Patch Changes
+
+-   Updated dependencies [9f3752d4]
+    -   @khanacademy/wonder-blocks-core@5.3.1
+
 ## 2.0.14
 
 ### Patch Changes