+++ packages/wonder-blocks-testing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @khanacademy/wonder-blocks-testing
 
+## 8.0.20
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-data@11.0.16
+
+## 8.0.19
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-data@11.0.15
+
 ## 8.0.18
 
 ### Patch Changes