+++ packages/wonder-blocks-tooltip/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @khanacademy/wonder-blocks-tooltip
 
+## 2.1.1
+
+### Patch Changes
+
+-   Updated dependencies [1344436f]
+    -   @khanacademy/wonder-blocks-core@5.4.0
+    -   @khanacademy/wonder-blocks-modal@4.0.18
+    -   @khanacademy/wonder-blocks-layout@2.0.16
+    -   @khanacademy/wonder-blocks-typography@2.1.2
+
+## 2.1.0
+
+### Minor Changes
+
+-   9f3752d4: Added custom styling for background color, text color, and padding to Tooltip
+
+### Patch Changes
+
+-   Updated dependencies [9f3752d4]
+    -   @khanacademy/wonder-blocks-typography@2.1.1
+    -   @khanacademy/wonder-blocks-core@5.3.1
+    -   @khanacademy/wonder-blocks-modal@4.0.17
+    -   @khanacademy/wonder-blocks-layout@2.0.15
+
 ## 2.0.16
 
 ### Patch Changes